Refactor data model
