+++ DasContract/DasContract/Abstraction.Interface/INamable.cs
@@ -6,6 +6,9 @@ namespace DasContract.DasContract.Abstraction.Interface
 {
     public interface INamable
     {
+        /// <summary>
+        /// Display (reader friendly) name of this entity
+        /// </summary>
         string Name { get; set; }
     }
 }
