+++ DasContract/DasContract/Abstraction/DataModel/Property/Reference/ReferenceContractPropertyType.cs
@@ -1,0 +1,8 @@
+﻿namespace DasContract.Abstraction.DataModel.Property.Reference
+{
+    public enum ReferenceContractPropertyType
+    {
+        Entity,
+        EntityCollection
+    }
+}
